+++ src/types/rtsp_version.c
@@ -4,10 +4,8 @@
 
 #include <assert.h>
 #include <inttypes.h>
-#include <stdio.h>
+#include <stdint.h>
 #include <string.h>
-
-#include <alloca.h>
 
 ssize_t SmolRTSP_RtspVersion_serialize(
     const SmolRTSP_RtspVersion *restrict self, SmolRTSP_Writer w) {
@@ -36,20 +34,20 @@ SmolRTSP_ParseResult SmolRTSP_RtspVersion_parse(
     MATCH(smolrtsp_match_numeric(input));
     minor = CharSlice99_from_ptrdiff(minor.ptr, input.ptr);
 
-    uint8_t major_int, minor_int;
+    uintmax_t major_int, minor_int;
 
-    if (sscanf(CharSlice99_alloca_c_str(major), "%" SCNu8, &major_int) != 1) {
+    if (!smolrtsp_parse_uint(major, UINT8_MAX, &major_int)) {
         return SmolRTSP_ParseResult_Failure(
             SmolRTSP_ParseError_TypeMismatch(SmolRTSP_ParseType_Int, major));
     }
 
-    if (sscanf(CharSlice99_alloca_c_str(minor), "%" SCNu8, &minor_int) != 1) {
+    if (!smolrtsp_parse_uint(minor, UINT8_MAX, &minor_int)) {
         return SmolRTSP_ParseResult_Failure(
             SmolRTSP_ParseError_TypeMismatch(SmolRTSP_ParseType_Int, minor));
     }
 
-    self->major = major_int;
-    self->minor = minor_int;
+    self->major = (uint8_t)major_int;
+    self->minor = (uint8_t)minor_int;
 
     return SmolRTSP_ParseResult_complete(input.ptr - backup.ptr);
 }